rearrange some classes
