+++ AvaGui/Models/VersionCheckBody.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace AvaGui.Models
+{
+	public class VersionCheckBody
+	{
+		[JsonPropertyName("tag_name")]
+		public string TagName { get; set; }
+	}
+}
